Plumb in and add tests for procession.log
procession.log.setup was not being called during WSGI app
creation, and there were no tests for the module. This patch
adds a call to procession.log.setup() before the construction
of the falcon.API() object and adds tests for validating the
conf_file and removal of null handlers from the root logger.