Remove deleted_at from all models
We were not using the deleted_at column, nor were we using a
"shadow" table model, so there's no real point in having the
column. We use real deletes, not fake deletes.

We plan to have an audit log table that stores create, delete, and
update audit records of all important models. Data can be restored
if needed from such a table manually.