Adds API resource routing for /organizations
Wire up the /organizations router in the API.

Also, to simplify testing, made some changes to the DB API that added
the commit kwarg to all delete methods and changed the default to be
True. This allowed me to take the session.commit() manual calls out of
the HTTP API resource object methods.