Add user group membership
Documents the DB API module, and adds functionality around user
group membership. In the process of writing this code, I realized
there was a problem with the cascading deletes of org -> group ->
ugm records, due to SQLAlchemy not cascading deletes if you execute
a DELETE against the underlying table instead of using the
session.delete() method. So, this patch also adds some code to ensure
that when an org is deleted, that groups and group membership are also
deleted.

Partially addresses Issue #32