Major refactoring around objects and DB API
Changes create methods in DB API for organization and group to use
objects -- which have validation built in to them. In the process, we
changed from UUID identifiers to integer identifiers, for scale and
performance reasons, as well as human-readability.

Moves objects out of procession/api/ and into the main procession/
namespace, and moves some of the serialization base routines out of the
api helpers and into the main procession helpers.