Limit camelCasing to just the cap'n'p files
It was really annoying to have to use camelCasing for attributes on the
object models just because cap'n'p forces you to use camelCasing, even
when Python's standard conventions are to use under_score_names. This
patch changes the base object model attributes to be under_score_names
and adds a thin translation layer for when the Cap'n'p message objects
need to be set/get.