Adds save() method implementation in store driver
Adds the ability to call procession.store.Store.save(), passing in a
single procession.objects.Object instance. The SQL store driver is
updated with a method that translates the object into a set of DB API
parameters to a DB API method to create or update a record in the DB.
Additional methods and parameters were added to the base
procession.objects.Object class to deal with determining if an object
has ever been saved to backing storage (is_new()) and to grab the set of
field names for any fields that have had their values changed.