Prevent overwrite of optional fields in user
Because of the way proto3 works, there is no way to distinguish between the
default value of a primitive field and that field not being set. So, there was
no way to determine from the original User message whether the display name or
email field had been set to the default value ("") or not, which meant the code
was just setting these fields to "" even if they were not specified on the
command line.

To address this issue, we bring in a wrappers.proto file and rework the
SetUserRequest message definition to include StringValue message fields instead
of string primitives. And we only construct these StringValue message fields if
the user specified a display name or email on the command line.