Simplify the UX for user get
Specifying --display-name, --email, --uuid, or --slug on the command
line is really quite a pain. So, in this patch, we change the way that
user records are looked up by allowing the user to specify a UUID, an
email, or just a regular old string on the command line:

p7n user get <search>

The <search> string is passed as-is to the backend IAM server, which
analyzes the search string to see if the search string looks like an
email or UUID and if so, grabs the user record for those unique fields.
If <search> does not look like either a UUID or email, then the WHERE
expression used to find the user record looks at the display_name and
slug fields with an OR condition.