Fixes #19: show >1 level of org treeview
This patch fixes Issue #19 by properly adding >1 level of nesting to the output
of `p7n organization list --tree`. We use a map of string -> []*orgTreeNode to
do a single pass through the returned list of organizations and look up a
parent org tree node if it has yet to be created by the time we hit on a node
that has it as its parent.