include predecessor members in child org members
We ensure that the user that created a new top-level root organization is
automatically added to the membership of the root organization.

We do not however, automatically add the creating user to the membership of a
child organization. This is because only users that have been added to an
organization's membership are allowed to create child organizations within that
organization (sub)tree, and since the creating user is already added to a
parent's membership, there's no need to add the user to the child's membership.

This is all good, but we need to ensure that users that are members of
ascendants in an org tree are also listed when looking at a child
organization's membership.

This patch adds the appropriate nested sets query constructs to select users
from the organization_users table that are in the target organization and any
of that organization's predecessors in the same org tree.

Closes #24