Implement phase one of user delete
Implements the delete of a user from Procession. We do some basic checking to
see if the user belongs to root organizations that have other users in them. If
so, we stop the deletion. Still need to add code that does deeper inspection
and ensures that we can't orphan an organization by deleting a user that is the
only member of a parent organization.

Addresses Issue #13