Complete implementation of user delete
In this patch, we clean up the logic of deleting a user record. We needed to
return an appropriate error message explaining why particular users cannot be
deleted if their deletion would result in the orphaning of a root organization.

Orphaning of an organization can occur if we delete a user that is the sole
member of a root organization and that root organization has child
organizations that have users associated with them. Upon deleting the user, we
would be in a situation where the root organization could not be modified or
deleted.

Closes Issue #13