begin removal of global state
Instead of using global state for logging in the RPC and DB layers, we
want to instead pass a pointer to a Context object that can be used to
store configuration information and loggers. This patch modifies the
calling interface for the iam/db routines to pass a context object that
itself contains a logger and a sql.DB pointer.