change archiver to eventlog
Changes the pkg/archive package to pkg/event to better represent what
the damn thing is. Changes the package interface from an Archive()
function to instead a NewLogger().Write() function that builds a
pb.Event message to write into the event log.

An archival utility will be built on top of this eventlog functionality.

Issue #14