continue refining the eventing interface
As expected, as I integrate the new event writing mechanism into Procession,
it's become evident that the interface needed to be more generic and
accomodating of multiple event streams. For instance, we'd like to be able to
both write to an event log as well as send out an event message to a
notification queue. So, this patch changes the public eventing interface to not
mention anything about logging but instead just be a generic Events struct with
a single Write method.