continue removing global state
This patch moves the pkg.iam.db.Context struct to a new package pkg/context
that is generic and can be used for more than just the IAM service. The end
goal is to get rid of global state, including global loggers. To do this
cleanly, we pass a pointer to the new pkg.context.Context struct that is
created in each service's main() function.