authz: add role DB schema and role get fn
Begins the process of adding in the role and role-permission storage. A new set
of tables in the IAM DB schema is added and the initial role_get IAM service
API RPC method is implemented.

Regardless of whether we rework the authz module to use Casbin after @hsluoyz's
excellent feedback and suggestions for how to do the type of policy enforcement
we need, we will still need a strategy for storing the permissions information
and role information in backend storage. Since we plan to use gRPC and
protobuffers for our API and data modeling, this step is essential even if we
go with Casbin.

Issue #11