authz: remove ScopedPermission enum type
Permissions are not scoped. Rather, the things to which a permission is
applied may be scoped. In this patch, we remove the duplicative
ScopedPermission enum and change the Permissions message to contain a
PermissionSet message and a map of string to PermissionSet messages,
one for system permissions the other for permissions scoped to a
particular organization.

Issue #11