Prevent duplicate display names for organizations
This patch prevents top-level (and within an organization) duplication
of display names by reworking the unique indexes on the organizations
table to be on the slug field and then auto-constructing child
organization slugs as the concatenation of the root organization slug
with the newly-supplied organization display name.

Attempting to create an organization or child organization with the
same name as another now results in an error.

Closes Issue #47