create Storage abstraction layer
Embeds a Storage struct as an abstraction layer over the data storage.
Also moves log and context to unexported attributes of the Server
struct. Following patches will remove the ctx attribute of the server
completely (this is what will be passed to various Server methods
instead of being a field in the struct.