pull more storage interface out into pkg/storage
Fully wraps the pkg/storage Storage struct with the IAMStorage struct,
which finishes the first major phase of abstraction of the storage
layer. Still need to figure out how to abstract Transaction and
Statement interfaces so that it's not dependent on concrete
database/sql structs, but this is getting us closer to the clean layer
we're looking for.