make gsr registry unexported member of iam.Server
Moves the final member of the pkg/iam/server:Server struct to be an
unexported member field. The gsr.Registry object was being set up
outside of pkg/iam/server:New() which meant it needed to be exported.
We undo that here and add a PROCESSION_SERVICE_NAME configuration
option and env variable to contain the needed configuration item for
setting up the gsr registry object inside of pkg/iam/server:New()