support pagination on non-unique sort fields
This is the first pass for supporting limit/marker pagination on
non-unique sort fields. This is working, but the code is a bit ugly so
followup patches will clean it up.

We add a couple new parameters to the pkg/sqlutil's AddMarkerWhere()
function -- a list of SortFieldInfo structs and a functor that can be
used, if needed, to get the value of the sort field if uuid is not used
for sorting.

Issue #109